memoize existing feed list to improve performance
